backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.6.0
python-dotenv==1.0.1
requests==2.32.3
google-generativeai==0.8.3
